auto_dec_ref_count_stack

+++ spvm_opcode.h
@@ -329,7 +329,7 @@ enum {
 extern const char* const SPVM_OPCODE_C_CODE_NAMES[];
 
 struct SPVM_opcode {
-  int32_t code;
+  uint16_t code;
   int32_t operand0;
   int32_t operand1;
   int32_t operand2;
